Move schemas to separate files
