+++ backend/lm_backend/api/cruds/generic.py
@@ -7,7 +7,7 @@ from sqlalchemy import Column, select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.sql.expression import UnaryExpression
 
-from lm_backend.api.schemas import BaseCreateSchema, BaseUpdateSchema
+from lm_backend.api.schemas.base import BaseCreateSchema, BaseUpdateSchema
 from lm_backend.database import Base, search_clause, sort_clause
 
 ModelType = TypeVar("ModelType", bound=Base)
