+++ backend/lm_backend/api/routes/bookings.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.booking import BookingCRUD
 from lm_backend.api.models.booking import Booking
-from lm_backend.api.schemas import BookingCreateSchema, BookingSchema, BookingUpdateSchema
+from lm_backend.api.schemas.booking import BookingCreateSchema, BookingSchema, BookingUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
 from lm_backend.session import get_session
