+++ backend/lm_backend/api/routes/clusters.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.generic import GenericCRUD
 from lm_backend.api.models.cluster import Cluster
-from lm_backend.api.schemas import ClusterCreateSchema, ClusterSchema, ClusterUpdateSchema
+from lm_backend.api.schemas.cluster import ClusterCreateSchema, ClusterSchema, ClusterUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
 from lm_backend.session import get_session
