+++ backend/lm_backend/api/routes/features.py
@@ -6,10 +6,12 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from lm_backend.api.cruds.generic import GenericCRUD
 from lm_backend.api.models.feature import Feature
 from lm_backend.api.models.inventory import Inventory
-from lm_backend.api.schemas import (
+from lm_backend.api.schemas.feature import (
     FeatureCreateSchema,
     FeatureSchema,
     FeatureUpdateSchema,
+)
+from lm_backend.api.schemas.inventory import (
     InventoryCreateSchema,
     InventoryUpdateSchema,
 )
