+++ backend/lm_backend/api/routes/jobs.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.generic import GenericCRUD
 from lm_backend.api.models.job import Job
-from lm_backend.api.schemas import JobCreateSchema, JobSchema, JobUpdateSchema
+from lm_backend.api.schemas.job import JobCreateSchema, JobSchema, JobUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
 from lm_backend.session import get_session
