+++ backend/lm_backend/api/routes/license_servers.py
@@ -6,8 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from lm_backend.api.constants import LicenseServerType
 from lm_backend.api.cruds.generic import GenericCRUD
 from lm_backend.api.models.license_server import LicenseServer
-from lm_backend.api.schemas import LicenseServerCreateSchema, LicenseServerSchema, LicenseServerUpdateSchema
-from lm_backend.api.constants import LicenseServerType
+from lm_backend.api.schemas.license_server import LicenseServerCreateSchema, LicenseServerSchema, LicenseServerUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
 from lm_backend.session import get_session
