+++ backend/lm_backend/api/routes/products.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from lm_backend.api.cruds.generic import GenericCRUD
 from lm_backend.api.models.product import Product
-from lm_backend.api.schemas import ProductCreateSchema, ProductSchema, ProductUpdateSchema
+from lm_backend.api.schemas.product import ProductCreateSchema, ProductSchema, ProductUpdateSchema
 from lm_backend.permissions import Permissions
 from lm_backend.security import guard
 from lm_backend.session import get_session
